# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transparência API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/blocos": {
            "get": {
                "tags": [
                    "/blocos"
                ],
                "summary": "listar blocos",
                "operationId": "b7e8c77ab80118075b10ac8f66d89b9b",
                "responses": {
                    "200": {
                        "description": "sucesso"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/blocos/centralidade/{centralidade}": {
            "post": {
                "tags": [
                    "/blocos"
                ],
                "summary": "Cadastrar um bloco numa centralidade",
                "operationId": "b8ef5a9a432325f903c6877bf37e9287",
                "parameters": [
                    {
                        "name": "centralidade",
                        "in": "path",
                        "description": "id da centralidade",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "c_descbloco": {
                                        "description": "denominação do bloco",
                                        "type": "string"
                                    },
                                    "c_ruabloco": {
                                        "description": "ruas que delimitam o bloco",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "bloco cadastrado com sucesso"
                    },
                    "412": {
                        "description": "Erro ao validar os dados"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/blocos/{bloco}": {
            "get": {
                "tags": [
                    "/blocos"
                ],
                "summary": "mostrar um bloco",
                "operationId": "c94deb4f7e6bf0a0f3476a6ce76e1279",
                "parameters": [
                    {
                        "name": "bloco",
                        "in": "path",
                        "description": "id do bloco",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sucesso"
                    },
                    "404": {
                        "description": "blocos não encontrado"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "/blocos"
                ],
                "summary": "Atualizar um bloco de uma centralidade",
                "operationId": "f948c9b3d404a04d4038063308055155",
                "parameters": [
                    {
                        "name": "bloco",
                        "in": "path",
                        "description": "id do bloco",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "c_descbloco": {
                                        "description": "denominação do bloco",
                                        "type": "string"
                                    },
                                    "c_ruabloco": {
                                        "description": "ruas que delimitam o bloco",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "bloco cadastrado com sucesso"
                    },
                    "412": {
                        "description": "Erro ao validar os dados"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "/blocos"
                ],
                "summary": "apagar um bloco",
                "operationId": "6c4f25766d9ed485f22b32b2e7b7f7dd",
                "parameters": [
                    {
                        "name": "bloco",
                        "in": "path",
                        "description": "id do bloco",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "bloco deletado com sucesso!"
                    },
                    "404": {
                        "description": "bloco não encontrado"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/caixas/{caixa}": {
            "get": {
                "tags": [
                    "/caixas"
                ],
                "summary": "mostrar um Taxa",
                "operationId": "8d1244c6fb474a7317824818c5158a64",
                "parameters": [
                    {
                        "name": "caixa",
                        "in": "path",
                        "description": "id do caixa",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sucesso"
                    },
                    "404": {
                        "description": "caixa não encontrado"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}